Create garbler.py
For imperfect translations - e.g. "his brings me to the last of the big questions: the future of the human race. If we are the only intelligent beings in the galaxy, we should make sure we survive and continue." -> "This brings me to the final of the big questions the future of the human race. If we are the only [unknown word] beings in the galaxy, we should spawn sure we survive and continue."